Adjust structures of modules
