+++ src/grammar/structures/mod.rs
@@ -6,7 +6,7 @@ mod save_frame;
 mod tag_value_line;
 
 pub use data_block::DataBlock;
-pub use data_items::DataItems;
+pub use data_items::{CIFDataType, DataItems};
 pub use loop_struct::LoopUnit;
 pub use save_frame::SaveFrame;
 pub use tag_value_line::SingleLineData;
