+++ src/grammar/tags_values/mod.rs
@@ -2,4 +2,4 @@ mod tag;
 mod values;
 
 pub use tag::Tag;
-pub use values::Value;
+pub use values::{CIFValue, Value};
